Refactor image APIs.
